+++ src/lib/query/tx.ts
@@ -85,3 +85,42 @@ export const getTxsCountByPoolId = graphql(`
     }
   }
 `);
+
+export const getTxs = graphql(`
+  query getTxs($offset: Int!, $pageSize: Int!) {
+    transactions(
+      order_by: { block_height: desc }
+      offset: $offset
+      limit: $pageSize
+    ) {
+      block {
+        height
+        timestamp
+      }
+      account {
+        address
+      }
+      hash
+      success
+      messages
+      is_clear_admin
+      is_execute
+      is_ibc
+      is_instantiate
+      is_migrate
+      is_send
+      is_store_code
+      is_update_admin
+    }
+  }
+`);
+
+export const getTxsCount = graphql(`
+  query getTxsCount {
+    transactions_aggregate {
+      aggregate {
+        count
+      }
+    }
+  }
+`);